# internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "VATUSAs internal API for use by other internal VATUSA services.",
        "title": "VATUSA Internal",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "VATUSA Support",
            "url": "http://www.swagger.io/support",
            "email": "vatusa6@vatusa.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "api.vatusa.local",
    "basePath": "/internal/v1",
    "paths": {
...
        "/disciplinary-log": {
            "get": {
                "description": "List all disciplinary log entries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplinary-log"
                ],
                "summary": "List all disciplinary log entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/disciplinary_log.Response"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new disciplinary log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplinary-log"
                ],
                "summary": "Create a new disciplinary log entry",
                "parameters": [
                    {
                        "description": "Disciplinary Log Entry",
                        "name": "disciplinary_log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/disciplinary_log.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/disciplinary_log.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/disciplinary-log/{id}": {
            "get": {
                "description": "Get a disciplinary log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplinary-log"
                ],
                "summary": "Get a disciplinary log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disciplinary Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/disciplinary_log.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a disciplinary log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplinary-log"
                ],
                "summary": "Update a disciplinary log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disciplinary Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Disciplinary Log Entry",
                        "name": "disciplinary_log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/disciplinary_log.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/disciplinary_log.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a disciplinary log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplinary-log"
                ],
                "summary": "Delete a disciplinary log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disciplinary Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Patch a disciplinary log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disciplinary-log"
                ],
                "summary": "Patch a disciplinary log entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disciplinary Log Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Disciplinary Log Entry",
                        "name": "disciplinary_log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/disciplinary_log.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/disciplinary_log.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "disciplinary_log.Request": {
            "type": "object",
            "required": [
                "cid",
                "entry"
            ],
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "entry": {
                    "type": "string",
                    "example": "Changed Preferred OIs to RP"
                },
                "vatusa_only": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "disciplinary_log.Response": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "'1234567' or 'System'"
                },
                "entry": {
                    "type": "string",
                    "example": "Changed Preferred OIs to RP"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "'1234567' or 'System'"
                },
                "vatusa_only": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "utils.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        }
    }
}